simplify init routines for bytes beginning work for removig lots of Byte access out of data and directly into Byte source add test harnesses for more ROM access

+++ Diz.Core/arch/CpuOperations.cs
@@ -6,7 +6,7 @@ namespace Diz.Core.arch
     public interface IReadOnlyCpuOperableByteSource
     {
         public Architecture GetArchitecture(int i);
-        int GetRomByte(int offset);
+        byte GetRomByte(int offset);
         public FlagType GetFlag(int i);
         int GetMxFlags(int i);
